requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
yfinance